# Remove reshape to speed up loading of scalar features in TensorFlow